// apps/desktop/src/common/templating/types.ts
type BinaryToTextEncoding = 'base64' | 'base64url' | 'hex' | 'binary';

import type {
  CloudProviderCredential,
  CookieJar,
  Environment,
  GrpcRequest,
  McpRequest,
  OAuth2Token,
  Project,
  Request,
  RequestGroup,
  Response,
  ResponseHeader,
  Services,
  SocketIORequest,
  UserUploadEnvironment,
  WebSocketRequest,
  Workspace,
} from 'insomnia-data';

type NodeCurlRequestType = Pick<Request, 'url' | 'method' | 'headers'> &
  Partial<Pick<Request, 'body' | 'authentication'>>;
export interface NodeCurlRequestOptions {
  request: NodeCurlRequestType;
  caCertficatePath?: string;
}
export interface NodeCurlResponseType {
  body: string;
  code: number;
  reason: string;
  status: string;
  responseTime: number;
  headers: ResponseHeader[];
  json: () => any;
  ok?: boolean;
}

export interface PluginStore {
  hasItem(arg0: string): Promise<boolean>;
  setItem(arg0: string, arg1: string): Promise<void>;
  getItem(arg0: string): Promise<string | null>;
  removeItem(arg0: string): Promise<void>;
  clear(): Promise<void>;
  all(): Promise<
    {
      key: string;
      value: string;
    }[]
  >;
}

export interface SerializedCookie {
  key?: string;
  value?: string;
  domain?: string;
  path?: string;
  expires?: string | Date;
  httpOnly?: boolean;
  secure?: boolean;
  hostOnly?: boolean;
}

export type RenderPurpose = 'send' | 'general' | 'preview' | 'script' | 'no-render';
// ...
export interface BaseRenderContext {
  getMeta: () => { requestId?: string; workspaceId?: string };
  getKeysContext: () => { keyContext: Record<string, string> }; // { keyContext: { 'env var name': 'Base Env' } };
  getPurpose: () => RenderPurpose | undefined;
  getExtraInfo: () => { requestChain: string[] } | undefined;
  getEnvironmentId: () => string | undefined;
  getGlobalEnvironmentId: () => string | undefined;
  getProjectId: () => string | undefined;
  [key: string]: any;
}

export interface AppPromptOptions {
  label?: string;
  defaultValue?: string;
  submitName?: string;
  inputType?: string;
}

export interface AppContext {
  alert: (title: string, message?: string) => Promise<void>;
  dialog: (
    title: string,
    body: HTMLElement,
    options?: { onHide?: () => void; tall?: boolean; skinny?: boolean; wide?: boolean },
  ) => Promise<void>;
  prompt: (title: string, options?: AppPromptOptions) => Promise<string>;
  getPath: (name: string) => Promise<string>;
  getInfo: () => { version: string; platform: NodeJS.Platform };
  showSaveDialog: (options?: { defaultPath?: string }) => Promise<string | null>;
  clipboard: { readText(): Promise<string>; writeText(text: string): Promise<void>; clear(): Promise<void> };
}
export interface PluginTemplateTagContext {
  app: AppContext;
  store: PluginStore;
  network: {
    sendRequest(
      request: Request,
      extraInfo?: {
        requestChain: string[];
        environmentId?: string;
      },
    ): Promise<Response>;
    sendRequestWithoutSideEffects(options: NodeCurlRequestOptions): Promise<NodeCurlResponseType>;
  };
  context: BaseRenderContext & {
    value: string | number;
  };
  meta: { requestId?: string; workspaceId?: string };
  renderPurpose?: RenderPurpose;
  util: {
    nodeOS: () => Promise<{
      arch: string;
      platform: NodeJS.Platform;
      release: string;
    }>;
    readFile: (path: string) => Promise<string>;
    decode: (buffer: Buffer, encoding?: string) => Promise<string>;
    encode: (input: string, encoding: BinaryToTextEncoding) => Promise<string>;
    render: (str: string) => string | Promise<string | null>;
    openInBrowser?: (url: string) => void;
    models: {
      request: {
        getById: (id: string) => Promise<Request | undefined>;
        getAncestors: (request: Request) => Promise<(Request | RequestGroup | Workspace)[]>;
      };
      cloudCredential: {
        getById: (id: string) => Promise<CloudProviderCredential | undefined>;
        update: (
          originCredential: CloudProviderCredential,
          patch: Partial<CloudProviderCredential>,
        ) => Promise<CloudProviderCredential>;
      };
      workspace: { getById: (id: string) => Promise<Workspace | undefined> };
      oAuth2Token: { getByRequestId: (id: string) => Promise<OAuth2Token | undefined> };
      cookieJar: {
        getOrCreateForParentId: (parentId: string) => Promise<CookieJar>;
        getCookiesForUrl: (parentId: string, url: string) => Promise<SerializedCookie[]>;
      };
      response: {
        getLatestForRequestId: Services['response']['getLatestForRequestId'];
        getBodyBuffer: Services['helpers']['getResponseBodyBuffer'];
      };
      settings: {
        get: Services['settings']['get'];
      };
    };
  };
}